Python

+++ Ejercicios condicionales/Ejercicio 34 UD1.py
@@ -1,5 +1,5 @@
 # Corrige los 4 errores o añade el código que necesites para que el siguiente programa se ejecute correctamente.
-número=float(input("introduce el número por favor: "))
+número=int(input("introduce el número por favor: "))
 if número%2==0:
     print("el número", número, "es par")
 else:
